Fix high severity CodeQL issue
`Comparison of narrow type with wide type in loop condition`

The type of `TOTAL_DRIVER_COUNT` is `int`.
The type of the loop variable was `uint8_t`.
As a result, a high severity alert is generated by CodeQL.

This fix does the following:
1. Fix the CodeQL alert by changing the loop variable type to use `int`.
2. Explicitly limiting the loop to MAXIMUM_TOTAL_DRIVER_COUNT.
3. Sets the originally named variable to the loop variable (but c-style cast to type `uint8_t`).

In addition, because there was not previously any code to ensure that
`TOTAL_DRIVER_COUNT` would be in the expected range of `[0x01..0xFF]`,
added a `TU_ASSERT()` at the appropriate location to ensure this condition.